@@ -4,60 +4,59 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "client": "webpack-dev-server --optimize-minimize --open --config webpack.dev.js",
+    "client": "webpack-dev-server --open --config webpack.dev.js",
     "server": "nodemon --exec babel-node src/server.js --ignore client",
     "start": "concurrently \"yarn server\" \"yarn client\"",
+    "start-prod": "concurrently \"yarn server\" \"npx serve dist/ -p 8080\"",
     "build": "concurrently \"webpack --config webpack.prod.js\" \"babel-node src/server.js\""
   },
   "author": "c4",
   "license": "MIT",
   "dependencies": {
-    "axios": "^0.18.0",
-    "concurrently": "^3.6.0",
-    "dotenv": "^6.0.0",
-    "express": "^4.16.3",
+    "axios": "^1.7.2",
+    "concurrently": "^8.2.2",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
     "kue": "^0.11.6",
-    "moment": "^2.22.2",
-    "node-raspistill": "0.0.15",
-    "onoff": "^3.1.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-toastify": "^4.1.0",
-    "redis": "^2.8.0",
-    "serialport": "^9.0.7",
+    "moment": "^2.30.1",
+    "node-raspistill": "2.0.1",
+    "onoff": "^6.0.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-toastify": "^10.0.5",
+    "redis": "^4.6.14",
+    "serialport": "^12.0.0",
     "slack-webhook": "^1.0.0",
-    "socket.io": "^2.1.1",
-    "socket.io-client": "^2.1.1"
+    "socket.io": "^4.7.5",
+    "socket.io-client": "^4.7.5"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^1.0.0",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
-    "lodash": "^4.17.10",
-    "node-sass": "^4.9.2",
-    "nodemon": "^1.18.1",
-    "sass-loader": "^7.0.3",
-    "style-loader": "^0.21.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.16.0",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-middleware": "^3.1.3",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.3"
+    "@babel/node": "^7.24.7",
+    "@babel/cli": "^7.24.7",
+    "@babel/core": "^7.24.7",
+    "babel-loader": "^9.1.3",
+    "@babel/preset-env": "^7.24.7",
+    "@babel/preset-react": "^7.24.7",
+    "@babel/preset-stage-0": "^7.8.3",
+    "css-loader": "^7.1.2",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.6.0",
+    "lodash": "^4.17.21",
+    "nodemon": "^3.1.3",
+    "sass": "^1.77.4",
+    "sass-loader": "^14.2.1",
+    "style-loader": "^4.0.0",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-middleware": "^7.2.1",
+    "webpack-dev-server": "^5.0.4",
+    "webpack-merge": "^5.10.0"
   },
   "babel": {
     "presets": [
-      "env",
-      "react"
-    ],
-    "plugins": [
-      "transform-object-rest-spread"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "prettier": {
