@@ -1,9 +1,11 @@
-import SerialPort from 'serialport'
+import { SerialPort, ReadlineParser } from 'serialport'
 
 export const parseData = data => {
-  return parseInt(data.replace(/[^0-9A-Fa-f]+/g, ''), 16)
+  return parseInt(data.replace(/[^0-9A-Fa-f]+/g, ''), 16).toString()
 }
 
-export default new SerialPort('/dev/ttyUSB0', { baudRate: 9600 }).pipe(
-  new SerialPort.parsers.Readline({ delimiter: '\n', encoding: 'ASCII' })
-)
+const port = new SerialPort({ path: '/dev/ttyUSB0', baudRate: 9600 })
+const parser = new ReadlineParser({ delimiter: '\n', encoding: 'ASCII' })
+const reader = port.pipe(parser)
+
+export default reader
